# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Log in with the admin password and receive a bearer token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResult"
        "401":
          description: Wrong password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
...
  /mgmtreport/comments/{id}:
    patch:
      operationId: updateMgmtreportComment
      tags: [mgmtreport]
      summary: Update a comment body (admin only)
      security:
        - adminBearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMgmtreportComment"
      responses:
        "200":
          description: Updated comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MgmtreportComment"
        "400":
          description: Invalid body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
        "401":
          description: Admin authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
        "404":
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
    delete:
      operationId: deleteMgmtreportComment
      tags: [mgmtreport]
      summary: Delete a comment (admin only)
      security:
        - adminBearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Admin authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
        "404":
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
...
  /mgmtreport/summary:
    get:
      operationId: getMgmtreportSummary
      tags: [mgmtreport]
      summary: Corporate P&L lines (수주~경상이익), plan vs actual monthly series
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Corporate P&L lines in 천 USD
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MgmtreportSummary"
        "404":
          description: No data for the requested year
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
  /mgmtreport/projects:
    get:
      operationId: listMgmtreportProjects
      tags: [mgmtreport]
      summary: Per-project plan vs actual monthly revenue/cogs series and annual outlook
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: includeGroups
          in: query
          required: false
          description: Include corporate rollup rows (DECV법인 취합본)
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Per-project series in 천 USD
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MgmtreportProjects"
        "404":
          description: No data for the requested year
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
  /mgmtreport/import/preview:
    post:
      operationId: previewMgmtreportImport
      tags: [mgmtreport]
      summary: Parse an uploaded 경영관리보고회 Excel and return a preview (no DB changes)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                year:
                  type: integer
              required: [file, year]
      responses:
        "200":
          description: Parsed preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MgmtreportImportPreview"
        "400":
          description: Bad upload (missing file, bad year, too large)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
        "422":
          description: Excel format error (Korean message)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
  /mgmtreport/import/apply:
    post:
      operationId: applyMgmtreportImport
      tags: [mgmtreport]
      summary: Parse an uploaded 경영관리보고회 Excel and replace mgmtreport data in the DB
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                year:
                  type: integer
              required: [file, year]
      responses:
        "200":
          description: Import applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MgmtreportImportResult"
        "400":
          description: Bad upload (missing file, bad year, too large)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
        "422":
          description: Excel format error (Korean message)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
  /mgmtreport/comments:
    get:
      operationId: listMgmtreportComments
      tags: [mgmtreport]
      summary: List saved 실적/전망 comments for a year+month
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: month
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
        - name: section
          in: query
          required: false
          description: Filter by section (analysis=실적, outlook=전망)
          schema:
            type: string
            enum: [analysis, outlook]
      responses:
        "200":
          description: Comments (newest first)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MgmtreportCommentList"
        "400":
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
    post:
      operationId: createMgmtreportComment
      tags: [mgmtreport]
      summary: Save a 실적/전망 comment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMgmtreportComment"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MgmtreportComment"
        "400":
          description: Invalid body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
...
components:
  securitySchemes:
    adminBearer:
      type: http
      scheme: bearer
  schemas:
    AdminLoginBody:
      type: object
      properties:
        password:
          type: string
          minLength: 1
      required: [password]
    AdminLoginResult:
      type: object
      properties:
        token:
          type: string
        expiresAt:
          type: string
          description: ISO timestamp when the token expires
      required: [token, expiresAt]
...
    UpdateMgmtreportComment:
      type: object
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 2000
      required: [body]
    MgmtreportComment:
      type: object
      properties:
        id:
          type: integer
        year:
          type: integer
        month:
          type: integer
        section:
          type: string
          enum: [analysis, outlook]
        body:
          type: string
        createdAt:
          type: string
      required: [id, year, month, section, body, createdAt]
    MgmtreportCommentList:
      type: object
      properties:
        comments:
          type: array
          items:
            $ref: "#/components/schemas/MgmtreportComment"
      required: [comments]
    CreateMgmtreportComment:
      type: object
      properties:
        year:
          type: integer
        month:
          type: integer
          minimum: 1
          maximum: 12
        section:
          type: string
          enum: [analysis, outlook]
        body:
          type: string
          minLength: 1
          maxLength: 2000
      required: [year, month, section, body]
    MgmtreportImportProjectPreview:
      type: object
      properties:
        name:
          type: string
        siteCode:
          type: [string, "null"]
        isGroup:
          type: boolean
        revenueActualTotal:
          type: number
        cogsActualTotal:
          type: number
        monthlyCount:
          type: integer
      required: [name, siteCode, isGroup, revenueActualTotal, cogsActualTotal, monthlyCount]
    MgmtreportImportPreview:
      type: object
      properties:
        year:
          type: integer
        unit:
          type: string
        projectCount:
          type: integer
        monthlyCount:
          type: integer
        annualCount:
          type: integer
        pnlCount:
          type: integer
        monthsWithActual:
          type: array
          items:
            type: integer
        totals:
          type: object
          properties:
            revenuePlan:
              type: number
            revenueActual:
              type: number
            cogsActual:
              type: number
          required: [revenuePlan, revenueActual, cogsActual]
        projects:
          type: array
          items:
            $ref: "#/components/schemas/MgmtreportImportProjectPreview"
      required:
        [year, unit, projectCount, monthlyCount, annualCount, pnlCount, monthsWithActual, totals, projects]
    MgmtreportImportResult:
      allOf:
        - $ref: "#/components/schemas/MgmtreportImportPreview"
        - type: object
          properties:
            applied:
              type: boolean
          required: [applied]
    ApiErrorMessage:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    MgmtreportLine:
      type: object
      properties:
        code:
          type: string
          description: Line code (new_orders, order_profit, revenue, cogs, gross_profit, sga, op_profit1, op_profit2, finance_cost, interest_income, ordinary_profit, ...)
        label:
          type: string
          description: Original Korean label
        plan:
          type: array
          description: 12 monthly plan values, index 0 = January
          items:
            type: number
        actual:
          type: array
          description: 12 monthly actual/forecast values, index 0 = January
          items:
            type: number
        planTotal:
          type: number
        actualTotal:
          type: number
      required: [code, label, plan, actual, planTotal, actualTotal]
    MgmtreportSummary:
      type: object
      properties:
        year:
          type: integer
        unit:
          type: string
          description: Amount unit (천 USD)
        lines:
          type: array
          items:
            $ref: "#/components/schemas/MgmtreportLine"
      required: [year, unit, lines]
    MgmtreportAnnualPoint:
      type: object
      properties:
        year:
          type: integer
        scenario:
          type: string
          description: actual (전년 실적) or forecast (전망)
        revenue:
          type: number
        cogs:
          type: number
      required: [year, scenario, revenue, cogs]
    MgmtreportProject:
      type: object
      properties:
        name:
          type: string
        siteCode:
          type: string
          nullable: true
        isGroup:
          type: boolean
          description: True for corporate rollup rows (DECV법인 취합본)
        revenuePlan:
          type: array
          items:
            type: number
        revenueActual:
          type: array
          items:
            type: number
        cogsPlan:
          type: array
          items:
            type: number
        cogsActual:
          type: array
          items:
            type: number
        annual:
          type: array
          items:
            $ref: "#/components/schemas/MgmtreportAnnualPoint"
      required: [name, isGroup, revenuePlan, revenueActual, cogsPlan, cogsActual, annual]
    MgmtreportProjects:
      type: object
      properties:
        year:
          type: integer
        unit:
          type: string
        projects:
          type: array
          items:
            $ref: "#/components/schemas/MgmtreportProject"
      required: [year, unit, projects]
...