# lib/api-zod/src/generated/types/adminLoginBody.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminLoginBody {
  /** @minLength 1 */
  password: string;
}