# lib/api-zod/src/generated/types/adminLoginResult.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminLoginResult {
  token: string;
  /** ISO timestamp when the token expires */
  expiresAt: string;
}