# lib/api-zod/src/generated/types/updateMgmtreportComment.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateMgmtreportComment {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  body: string;
}